+++ app/api/event-delete/route.ts
@@ -1,16 +1,26 @@
+import { auth } from '@/auth';
 import { db } from '@/db';
 import { events } from '@/db/schema';
 import { eq } from 'drizzle-orm';
-import { NextResponse, type NextRequest } from 'next/server';
+import { NextResponse } from 'next/server';
 
-export async function POST(req: NextRequest) {
+export const POST = auth(async (req) => {
   try {
+    const admin = req.auth;
+    if (!admin) {
+      throw new Error('not authenticated');
+    }
+    if (admin.user.role === 'USER') {
+      throw new Error('not authenticated');
+    }
+
     const res = await req.json();
 
     await db.delete(events).where(eq(events.id, res.id));
 
     return NextResponse.json({ message: 'success' });
-  } catch {
+  } catch (err) {
+    console.log(err);
     return NextResponse.json({ message: 'err' });
   }
-}
+});
